+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.0.3] - 2023-04-14
 ### Fixed
 - Map empty objects in result and input to custom types with noop field (`EmptyObject`, `EmptyInputObject`) - [#52](https://github.com/superfaceai/one-service/pull/52)
 
@@ -103,7 +105,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.0-rc.0 - 2022-01-11
 
-[Unreleased]: https://github.com/superfaceai/one-service/compare/v3.0.2...HEAD
+[Unreleased]: https://github.com/superfaceai/one-service/compare/v3.0.3...HEAD
+[3.0.3]: https://github.com/superfaceai/one-service/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/superfaceai/one-service/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/superfaceai/one-service/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/superfaceai/one-service/compare/v2.2.3...v3.0.0